Start User login

+++ src/main/java/com/moviezone/dai_api/model/dao/IUserDAO.java
@@ -3,7 +3,7 @@ package com.moviezone.dai_api.model.dao;
 import com.moviezone.dai_api.model.entity.User;
 
 public interface IUserDAO {
-    public User findUserById(int userId);
+    public User findUserById(long userId);
     public void createUser(User user);
-    public void deleteUser(int userId);
+    public void deleteUser(long userId);
 }
