+++ src/main/java/com/moviezone/dai_api/model/dao/UserDAOImplementation.java
@@ -17,7 +17,7 @@ public class UserDAOImplementation implements IUserDAO{
 
     @Override
     @Transactional(readOnly = true)
-    public User findUserById(int userId) {
+    public User findUserById(long userId) {
         Session currentSession = entityManager.unwrap(Session.class);
 
         User user = currentSession.get(User.class, userId);
@@ -35,7 +35,7 @@ public class UserDAOImplementation implements IUserDAO{
 
     @Override
     @Transactional
-    public void deleteUser(int userId) {
+    public void deleteUser(long userId) {
         Session currentSession = entityManager.unwrap(Session.class);
 
         Query theQuery = currentSession.createQuery("delete from User where id=:userId");
